Rename server tests to utility tests
